pyproject.toml
[tool.poetry]
name = "gjdanawa-uploader"
version = "0.1.0"
description = "gjdanawa-uploader"
authors = ["alchemine <djyoon0223@gmail.com>"]
readme = "README.md"
packages = [{include = "gjdanawa_uploader"}]

[tool.poetry.dependencies]
python = "^3.11"
dask = "^2024.7.0"
pyyaml = "^6.0.1"
pandas = "^2.2.2"
numpy = "^2.0.0"
matplotlib = "^3.9.1"
tabulate = "^0.9.0"
selenium = "^4.23.1"
easydict = "^1.13"
streamlit = "^1.37.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
